MLK-23352 soc: imx: scu: Fix corruption of header
The header of the message to send can be changed if the
response is longer than the request:
- 1st word, the header is sent
- the remaining words of the message are sent
- the response is received asynchronously during the execution of the
loop, changing the size field in the header
- the for loop test the termination condition using
the corrupted header

It is the case for the API build_info which has just a
header as request but 3 words in response.

This issue is loading message size before send.

This can be easily reproduced by adding udelay(1000) at the end of
MU_SendMessage

Fixes: 20c60c165929 ("MLK-21078-3 soc: imx: enable RX interrupt for IPC response")
Change-Id: I95431485ebb6ef8e883e4a1c7a2fceb36016c6fc
Signed-off-by: Franck LENORMAND <franck.lenormand@nxp.com>
Signed-off-by: Leonard Crestez <leonard.crestez@nxp.com>
Acked-by: Nitin Garg <nitin.garg@nxp.com>