MLK-23396: crypto: caam - ensure CAAM context is not sharing the cacheline
In caam_hash_state struct, caam_ctx buffer needs to have a
separate cacheline, not sharing it with "update" callback.
On imx8, the cacheline size is 64 and the MAX_CTX_LEN,
from caam_ctx buffer, is (8 + 64).
Therefore, add a ____cacheline_aligned to the update callback,
in caam_hash_state struct, to ensure that caam_ctx buffer is
not sharing the cacheline.

Signed-off-by: Iuliana Prodan <iuliana.prodan@nxp.com>
Reviewed-by: Horia Geantă <horia.geanta@nxp.com>