MA-17474 cpufreq: schedutil: Add freq_coefficient parameter for schedutil
When the schedutil, it will call map_util_freq() to
calculate the target frequence level. It will use below to calculate:

    next_freq = C * curr_freq * util_raw / max

By default the C = 1.25 in map_util_freq(). But in some performance
sensitive product, we want CPU working in higher freq for some low util
task to get better performance, so that add a freq_coefficient to
tuning the schedutil. The default freq_coefficient is 1. Then the
next_freq will be calculated by:

    next_freq = freq_coefficient * C * curr_freq * util_raw / max

TEST: $echo 2 > /sys/devices/system/cpu/cpufreq/schedutil/freq_coefficient
      $cat /sys/devices/system/cpu/cpufreq/schedutil/freq_coefficient

Change-Id: I6b3169c00c2c60c40c645de487b46fe90f655a3f
Signed-off-by: Haoran.Wang <elven.wang@nxp.com>