MLK-24753-2: ASoC: fsl_esai_mix: get pointer from dma driver directly
Previously update pointer in mix function according to dma callback,
which is not accurate for the callback maybe triggerred very late, that
cause the read pointer and write pointer point to same period, then
there is noise in output.
So get pointer directly from dma driver to avoid such issue.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Reviewed-by: Peng Zhang <peng.zhang_8@nxp.com>