MA-17820-1 set usermodehelper path to pass a cts case
A cts case "testConfigDisableUsermodehelper" requires enablement of
static usermodehelper, and the usermodehelper path need to be empty
or started one of "/vendor/", "/system_ext/" and "/system/".

With platform selinux rules, kernel can not execute binary from
"/vendor", set it to be "/system/bin/sh" then it can be used to start
other programs.

Change-Id: Iedaf4fa5f13ab288f85346abfa800a4236124a91
Signed-off-by: faqiang.zhu <faqiang.zhu@nxp.com>