MA-18695-2 [Android] Fix GKI compatibility issue about OF_DYNAMIC
if drm bridge(it6263/adv7511) probe failed(it may not plug in) and the drm
encoder(sec_mipi_dsim-imx/imx8mp-ldb) will bind failed. If one component bind failed,
all later component failed. system will bind the component continuely.

The key point is that the drm encoder do not know the bridge driver still not probe or
the bridge is not connected, they both return EPROBE_DEFER.(reference drm_of_find_panel_or_bridge)

The solution is that:
If it's the bridge driver do not probed in time, retry 4 times to wait the drm bridge driver ready;
Fake that bind successfully after retry 4 time to bind the drm bridge in case the bridge is
not connected

Change-Id: Iacbf0948bf77a33e3c32cf3f5ece745bc089d254
Signed-off-by: zhang sanshan <pete.zhang@nxp.com>