Revert "MA-18848 defconfig: Set CONFIG_FORCE_MAX_ZONEORDER to 14"
This reverts commit 6593ff381b9fac16af78e1f9810aef43b085d7a6.

This patch will break KMI, so revert it.

In include/linux/mmzone.h, the struct zone has one member defined
as "struct free_area free_area[MAX_ORDER]", and the MAX_ORDER
is defined as CONFIG_FORCE_MAX_ZONEORDER.
At default, the MAX_ORDER is 11, but with this patch,
it is 14, so the symbol of "struct zone" changes.

Change-Id: I1758615be2b8c413a2a7df7a4ca5b30b34944312
Signed-off-by: Jindong Yue <jindong.yue@nxp.com>