MA-18912[Android]imx8mq: make out fence been signaled when framebuffer off-screen
The out fence in vblank event is signaled when framebuffer start to scan out
to screen. So out fence is signaled when framebuffer in use. It cause android
framework reuse this buffer when it's busy and the screen is tearing.
Extract the fence information from the original vblank event and create a new
event to contain such fence. This new event can be signaled one vblank count
later(The framebuffer is off-screen at that time).

Change-Id: I12bce64bec8208bed2e3a2542913d5d58e4b8cc6
Signed-off-by: Zhang Bo <bo.zhang@nxp.com>