MA-19664 clks:Set the uart clock used by the console to critical
Because the earlycon of 8mm, 8mn and 8mq is open, they use the
corresponding uart clock. In this patch:

MA-19859 Gate off core and bus clock slice

We gate off the peripheral clock when registering. Because earlycon is
using uart clock at this time, we forcibly turn off uart clock, which
will cause system block. Therefore, set the corresponding uart
clock to critical and do not turn off this clock.

Change-Id: I636ab7c9f9c0d3c67d7f04b16b0107ce61d0e49e
Signed-off-by: Zhipeng Wang <zhipeng.wang_1@nxp.com>