MA-19449-2 Port the operations of decoder related VPU regs to TEE on imx8mq.
For the use of registers, the decoder are all passed to TEE through smc.
Use trusty_vpu_write/read to wrap the smc interface and iowrite32/ioread32,
for sensitive registers such as input and output,we use the OPT_SECURE_WRITE
smc type to store the buffer address in the TEE cache.When the decoding is
started, the VPU is turned on through the OPT_SECURE_CTRL_WRITE type
smc interface.

Conflicts:
	drivers/mxc/hantro/hantrodec.c

Change-Id: I06ee0cb493d2620204e9b3fdc08d636bff920f83
Signed-off-by: zhai.he <zhai.he@nxp.com>