MA-20199 Make the trusty core driver probe later
With the pkvm support, the FF-A share instructions will trap to
the FF-A handler in EL2 before they are routed to EL3. But as
the trusty core driver (which will issue SMC_FCZ_FFA_RXTX_MAP)
probes at subsys_initcall() which is early than the kvm driver
probe, which causes the mailbox mapping in kvm failed and all
following memory share failed.

This issue will impact platforms without GKI support, this commit
moves the trusty core driver probe to module_init() so it can
probe when kvm drivers ready.

Change-Id: I4182eadf879b5cb3819cbbf816f98a7c79b2b45b
Signed-off-by: Ji Luo <ji.luo@nxp.com>