LF-6002-3 i2c: imx-lpi2c: add switch to deal with the result of dma xfer
There is a special case that can cause problems.
When the message num that needs to be sent by DMA in an I2C message queue
is greater than 1, the previous code will cause only one message to be
sent and then the sending will be ended.

Add a switch to handle the result of dma xfer.

Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Reviewed-by: Haibo Chen <haibo.chen@nxp.com>