ANDROID: Use kthread for trusty work instead of wq
Per CPU, create a kthread with a wait queue to pause it when
there is no work to do. Instead of adding work to work queue,
wake the wait queue for the current cpu so that the work can
be completed.

For now, the thread priority follows the use_high_wq flag as
before: either normal or elevated.

Bug: 260139140
Signed-off-by: Brandon Anderson <bander9289@gmail.com>
Change-Id: I913c64d598f87e7c005f6702714c599da5a06da2