MA-21429-1 arm64: dts: imx8q: change USB Type-C port default power role to sink
This change does not affect the functionality of the
typec, it only changes the default power role.

The default power role is changed in order to pass
the following vts:
VtsHalUsbGadgetV1_1HostTest

The test has the following actions:
echo host > sys/class/typec/port0/data_role

If the current port role (sink) is not equal to the
default port role (source), the system will reset the
port and the adb device will be out of connection for
about 3s. vts reports an error: No test results.

If change the default port role to the same as the
current port role, the system will not reset the port.
vst passes.

Change-Id: I7f6e055d255353afd293db43b8d086f659a2e398
Signed-off-by: Zhipeng Wang <zhipeng.wang_1@nxp.com>