MA-21531 dts: arm64: 8mic: Remove duplicated interrupts node
The gpio-keys driver will constantly generate errors:
  irq: no irq domain found for gpio@21 !
before its interrupt parent pca9555 probed.
The issue happens when gpio-keys driver is built-in while pca9555
driver is built as a module.

It is not necessary to keep both "gpios" and "interrupts" in dts
because the irq can be get from "gpios" node with gpiod_to_irq().

Remove the "interrupts" node to eliminate the error logs.

Change-Id: I165b1944eb72781b99ff7be31b903ee57ae88359
Signed-off-by: Jindong Yue <jindong.yue@nxp.com>