MA-22309 display is not back after resume on evk_95
Android will do DPMS=off operation when suspend and DPMS=on when resume.
When DRM resume, the crtc state enable change: 0->1, and the mode_changed flag
will be set and mode_set() will be called. Then Android set DPMS=on operation
to enable display. It will set the crtc state active change: 0->1, it will also
set mode_changed and mode_set() called.
So the mode_set() is double called in display resume operation. The mipi dsi
phy will init two times and the second time will fail because the mux of phy
is locked.
Not return directly when mux is locked to make phy init continue to fix this
issue.

Change-Id: Iae58c1aa53d42e136fc8d6fe912f187b867832eb
Signed-off-by: Zhang Bo <bo.zhang@nxp.com>