MA-22361-2 Enable vpu vm1 for secure video on imx95
Enable vpu1 device for secure decoder.
Add trusty node for vpu1 device in dts. Based on the "trusty", vpu1
device will be registered with a different name to distinguish it
from the normal decoder. Besides, in order for the upper layer
to open the correct video device, use the driver field of the
capabilities of the secure decoder.

Change-Id: Ib92dd40131c7feb007c8f3bde4490cfe6487f960
Signed-off-by: He Zhai <zhai.he@nxp.com>