MA-22765 fix kernel dump in sysrq reboot
When the system reboot was triggered from some emergency
reboot like sysrq, the "cmd" can be passed as NULL pointer.

And because the reboot was triggered by kernel directly,
it won't follow the userspace reboot process to clean up
the resources in sequence, which may cause problem in some
cases like the communication with ELE(the AP core reboot
but ELE won't reboot). This commit will trigger a board
reset in such case to reload the whole system.

Change-Id: Ief4fc73161e0275fe43367a37e5ef2a4d849caea
Signed-off-by: Ji Luo <ji.luo@nxp.com>