LF-14281: media: amphion: Support dmabuf and v4l2 buffer without binding
When using VB2_DMABUF, the relationship between dma-buf and v4l2 buffer
may not one-to-one, a single dma-buf may be queued via different
v4l2 buffers, and different dma-bufs may be queued via the same
v4l2 buffer, so it's not appropriate to use the v4l2 buffer index
as the frame store id.

We can generate a frame store id according to the dma address.
Then for a given dma-buf, the id is fixed.

Driver now manages the frame store and vb2-buffer states independently.

When a dmabuf is queued via another v4l2 buffer before the buffer is
released by firmware, need to pend it until firmware release it.

Tested-by: xiahong.bao <xiahong.bao@nxp.com>
Reviewed-by: TaoJiang <tao.jiang_2@nxp.com>
Change-Id: Ic2d610b76b4ee1f642ec4f147697e2d06e3de3a2
Signed-off-by: Ming Qian <ming.qian@oss.nxp.com>