MA-23263 wave6: Allow 2x2 cropping size for the encoder.
The driver allows a 2x2 alignment crop size,
which does not adhere to the Wave6 hardware limitations.
However, the driver still encodes the crop size
according to the Wave6 hardware limitations.
Instead, it includes the conf_win info in the SPS
to indicate the exact crop size.
Driver should report the exact crop size in CAPTURE buffer.

Change-Id: I3978da4a013ea9fbf87865a51f1cd765f5532f58
Signed-off-by: Nas Chung <nas.chung@chipsnmedia.com>