MA-23391 [IMX8Q] VPU/DPU driver probe requires trusty driver probe complete
When all drivers are built in, the trusty driver will probe after
the vpu/dpu driver probe, which will result in a trusty dump. To solve
this issue, the vpu/dpu driver needs to determine whether the trusty
driver probe successfully before calling the trusty api, otherwise the
probe will be deferred.

Change-Id: I2cd20424e9c70dfff478096b28dfef2ca6eeb816
Signed-off-by: He Zhai <zhai.he@nxp.com>