LF-15200: mxc: vpu: wave6: Fix an rgb encoding failure issue
After update to driver v1.3.12, the encoding will fail, when the source
format is rgb.

The reason is the type coefficient variables is u32, but in
wave6_set_csc(), some negative value is set.  but in
wave6_check_enc_param(), it will return error if the u32 value is great
than 1023. then the negative u32 coefficient will always fail the
wave6_check_enc_param().

In wave6_check_enc_param(), convert the coefficient to a positive value,
then check its range.

Change-Id: Icf5ec63f53a7196db632bc184b03b170235088a5
Signed-off-by: Ming Qian <ming.qian@nxp.com>