MA-23963 Revert "LF-12944 firmware: arm_scmi: bus: bypass set fwnode for scmi cpufreq"
The reason for reverting the patch is that it bypasses setting the fwnode
for the SCMI cpufreq device, which conflicts with the following patch:

77ff6aec7c31 cpufreq: scmi: Skip SCMI devices that aren't used by the CPUs

Resulting in the cpufreq driver not being registered. Therefore, the
patch is reverted.

Reverting the patch introduces a potential risk:
If cpufreq.off=1 is set, it may cause GPU/VPU probe failures.

This reverts commit 2e8a946c6ee2f0d4c71f328299c4dd961ae605bb.

Change-Id: I6e56e23af6405fece51fdfa617c5bc6e7c58e7c7