MA-21429 arm64: dts: imx8q: change USB Type-C port to DRP (Dual-Role Power)
Currently, USB Type-C port is configured as a source-only port with DRD
(Dual-Role Data). However, this will bring issues if the USB port want to
ack as device mode when it's attached to non Type-C port, because the tcpm
(Type-C Manager) will stop CC toggling for source-only port. Actually,
this port is PD-capable and it can be configured as DRP too. In this way,
the CC pins will continue to toggle for different cases.
In this patch, we firstly change power-role to "dual" and then set
try-power-role to "source" based on original usecase. Since imx8q-mek
is powered by external power supply, "self-powered" is also added. Because
imx8q-mek doesn't draw current from vbus, we also add 5V/0A sink-pdo and
set op-sink-microwatt to 0mW.

Change-Id: I681cadeff4affee81dedb2bcf86edbc4b2e516ae
Signed-off-by: Zhipeng Wang <zhipeng.wang_1@nxp.com>