arm64: dts: imx952-evk: add ENETC sharing use case support
For the ENETC sharing use case of i.MX952-EVK, only the VF of ENETC 0 is
owned by A-core, and all the other resources of NETC are owned by M-core,
including the PCIe ECAM space. Therefore, A-core cannot directly access
the PCIe ECAM space, but the Linux PCIe driver needs to perform
enumeration to discover the PCIe functions and then probe the VF driver.
So the M-core create a 1MB shared memory as the virtual PCIe ECAM space,
so that Linux can directly read this virtual ECAM to perform enumeration.

In addition, since the I2C6 is used to reset the Ethernet PHYs and
control ENET2 IO daughter card, so this I2C bus is also controlled by
M-core. Currently, we just simply disable it in the DTS for this sharing
use case, and all devices which depend on this bus will be unavailabe on
Linux side. Users can use I2C RPMSG to make them available on Linux side
if they really need these devices. And netcmix_blk_ctrl is also disabled
due to it is owned by M-core.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>