MA-24293 phy: freescale: imx8m-pcie: Convert to module_param for modular build
Replace __setup() with module_param() to fix unused function warning
when the driver is built as a module.

The __setup() mechanism only works when the driver is built into the
kernel, causing imx8_pcie_phy_fine_tune() to become an unused function
when compiled as a module, which triggers a compilation error with
-Werror,-Wunused-function.

Using module_param() allows the parameter to work in both built-in and
modular configurations. The parameter can now be set via:
- Module load: modprobe phy-fsl-imx8m-pcie imx8_pcie_phy_tuned=1
- Kernel cmdline: phy-fsl-imx8m-pcie.imx8_pcie_phy_tuned=1

This maintains the same functionality for enabling PCIe PHY fine tuning
in i.MX8MP EP/RC validation systems to achieve Gen3 link speeds.

Change-Id: I3ff490784ba82c9f1c44b12abdd95ec92046f996
Signed-off-by: Zhipeng Wang <zhipeng.wang_1@nxp.com>