LF-17090-9 drm/bridge: synopsys: dw-mipi-dsi2: Support getting next bridge when attaching bridge
Since ADV7535 driver registers DSI device in drm_bridge_funcs->attach()
callback and this driver gets the next bridge and adds DRM bridge in
mipi_dsi_host_ops->attach(), the bridge chain would not be able to form
due to the "chicken or the egg" dilemma.  To solve it, this driver has
to support getting next bridge in this driver's drm_bridge_funcs->attach()
callback and add the DRM bridge at probe stage.

Reviewed-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Signed-off-by: Liu Ying <victor.liu@nxp.com>