LF-16840: mxc: vpu: wave511: Avoid hang due to inaccurate queued count
Vpu driver increase the queued count after send DEC_PIC cmd to firmware,
but if the cpu loading is very high, the decode done irq may be
triggered before that, then the queued count will be inaccurate,
which will prevent to flush the last severl display buffers.

It cause the pipeline hang as the EOS is not received.

Driver should increase the queued count before send DEC_PIC command,
then it won't be affacted by the cpu loading.

Signed-off-by: Ming Qian <ming.qian@nxp.com>
Reviewed-by: Zhou Peng <eagle.zhou@nxp.com>